pyproject.toml
[tool.poetry]
name = "adaptive-learning-engine"
version = "1.0.0"
description = "K1-K8 adaptive assessments powered by Neo4j Knowledge Graph + BKT + Gemini AI"
authors = ["Esen Dashnyam <esen.dashnyam@gmail.com>"]
readme = "README.md"
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
loguru = "^0.7.2"
neo4j = "^5.20.0"
google-generativeai = "^0.7.0"
python-multipart = "^0.0.9"

# LangGraph agentic loop
langgraph = "^0.3.0"
langchain-core = "^0.3.0"
langchain-google-genai = "^2.0.0"

# PostgreSQL (async)
sqlalchemy = { extras = ["asyncio"], version = "^2.0.0" }
asyncpg = "^0.30.0"
alembic = "^1.13.0"
greenlet = "^3.0.0"

# PDF ingestion
pymupdf = "^1.24.0"

# Knowledge graph construction
networkx = "^3.3"
yake = "^0.4.8"

# Web scraping / book ingestion
requests = "^2.32.0"
beautifulsoup4 = "^4.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
